autotag/autotaggers/rrj/requirements.txt
torch
torchvision
timm
pillow
safetensors
pydantic==2.10.6
msgspec